refactor: moving the MSI client
