+++ azurerm/internal/services/msi/client/client.go
@@ -1,4 +1,4 @@
-package msi
+package client
 
 import (
 	"github.com/Azure/azure-sdk-for-go/services/preview/msi/mgmt/2015-08-31-preview/msi"
@@ -9,7 +9,7 @@ type Client struct {
 	UserAssignedIdentitiesClient *msi.UserAssignedIdentitiesClient
 }
 
-func BuildClient(o *common.ClientOptions) *Client {
+func NewClient(o *common.ClientOptions) *Client {
 	UserAssignedIdentitiesClient := msi.NewUserAssignedIdentitiesClientWithBaseURI(o.ResourceManagerEndpoint, o.SubscriptionId)
 	o.ConfigureClient(&UserAssignedIdentitiesClient.Client, o.ResourceManagerAuthorizer)
